Remove implicit toService in favor of ToService case class and Endpoint.toService
